Update SpaceShip.java
This update changed how users control the movement of the spaceship from using a keyboard to a mouse. User could move the spaceship by dragging it around the frame ( We've added handling so that the spaceship won't be lost from your sight). To attack the aliens, users could fire missiles by pressing space key from keyboard.